requirements.txt
Django==1.8
django-material==0.6.0
PyMySQL==0.7.2
